refact : DelimiterAndBodyInfo(Dto)를 inner class -> top-level로 변경
