+++ src/main/java/stringCalc/DelimiterAndBodyInfo.java
@@ -1,0 +1,19 @@
+package stringCalc;
+
+public class DelimiterAndBodyInfo {
+    private String customDelimiter;
+    private String body;
+
+    public DelimiterAndBodyInfo(String delimiter, String body) {
+        this.customDelimiter = delimiter;
+        this.body = body;
+    }
+
+    public String getDelimiter(){
+        return customDelimiter;
+    }
+
+    public String getBody(){
+        return body;
+    }
+}
