+++ src/test/java/CustomDelimiterAndBodyExtractorTest.java
@@ -2,6 +2,7 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 import stringCalc.CustomDelimiterAndBodyExtractor;
+import stringCalc.DelimiterAndBodyInfo;
 
 import java.util.stream.Stream;
 
@@ -14,7 +15,7 @@ public class CustomDelimiterAndBodyExtractorTest {
     @ParameterizedTest
     @MethodSource("inputStringArguments")
     void givenMatchedPattern_whenExtract_thenGiveInputInfo(String input, String expected_body, String expected_del) {
-        CustomDelimiterAndBodyExtractor.DelimeterAndBodyInfo actual = CustomDelimiterAndBodyExtractor.extract(input);
+        DelimiterAndBodyInfo actual = CustomDelimiterAndBodyExtractor.extract(input);
         assertThat(actual).extracting("customDelimiter", "body").containsExactly(expected_del, expected_body);
     }
